Alerts grid now passes alerts data back to main page
Alerts data is held in two hash maps that map WeatherType to boolean.

There are two maps, one for if the user wants to see the alert, and one for if it is a priority alert.

It only maps the weather Types seen in the alerts pane.

These are passed back via the init() method in Main Page.
there is a flag that is set to true if the data is coming form the alerts tab, in which case it loads the alerts data, otherwise it loads settings data (the only other tab that calls that method)

To do this, I have had to also add fields to sceneResource, MainPage, and AlertsSettingsGridController