Creating persistent storage and updating settings
I have created a package inside frontend that holds classes for storing data persistently, allowing the settings (and in future the alerts) to work properly

The settings currently store things like location and windspeed scale permanently (and other things)

I have edited main page to load this information from persistent storage, and set the location API using this.

Also, if the user has sepected fahrenheit or kph, it displays these scales instead of celcius or mph and converts the windspeed and temperature values accordingly

I have also,
fixed bugs in the settings fxml file after the styling
Styled the location pane in a similar fashion

Done some code cleanup to remove redundant code from the old SceneResource mechanism, this placeholder class can eventually be delted I imagine

please test to see if the persistent storage works correctly, easist is to try London as a location, see if it stores it, then try Moskow and see if the temperature changes, (which it does on mine). But when you press back from settings, pause and wait for it to load, because loading the API info seems to take a while