Storing and Styling Alerts
(IMPORTANT NOTE AT BOTTOM)
The Alerts Pane has been styled in line with the long view styling and the main page

Now the Storage Classes store the alerts data as well, and the users prefernce is loaded when you go to the alerts tab.

This data can be fetched at any point in the application, for whoever implements the alerts.

The alerts information is now returned as two arrays of booleans, for the five weather types in the alerts pane. True means the radio button is selected, False means it is not.

IMPORTANT NOTE:
When testing you NEED to delete the file storage.xml in your local directory (unless you copy the one I have pushed)
Then when you run the program, the code will detect that it is missing and build a new one with a default setting for the alerts data in it
Test that you can edit the alerts radio buttons, press back, quit the program, re-load and see that the selection you made is retained.